Moves `replica` to `aws_dynamodb_table`.
